+++ tests/test_api_main.py
@@ -1,0 +1,7 @@
+from fastapi.testclient import TestClient
+
+
+def test_root(client: TestClient):
+    response = client.get("/")
+    assert response.status_code == 200
+    assert response.json() == {"message": "Finance Tracker is alive!"}
